REF: Rename package and update imports to obsidian-daylio-parser
This avoids the confusion caused by inconsistent naming: sometimes it was daylio_to_md, sometimes obsidian-daylio-parser, and sometimes daylio-obsidian-parser.
