+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "daylio-obsidian-parser"
+name = "obsidian-daylio-parser"
 dynamic = ["version", "readme"]
 requires-python = ">= 3.10"
 authors = [
@@ -19,9 +19,9 @@ classifiers = [
 keywords = [ "csv", "markdown", ".md", ".csv", "daylio", "obsidian", "file converter" ]
 
 [project.scripts]
-# Make daylio_to_md command be recognised by shell as a program
+# Make obsidian-daylio-parser command be recognised by shell as a program
 # No need to type: python -m <package> anymore
-daylio_to_md = "daylio_to_md.__main__:main"
+obsidian-daylio-parser = "obsidian_daylio_parser.__main__:main"
 
 [build-system]
 requires = ["setuptools", "build", "wheel", "versioningit"]
