+++ src/obsidian_daylio_parser/journal_entry.py
@@ -16,9 +16,9 @@ import typing
 import datetime
 from dataclasses import dataclass
 
-from daylio_to_md.config import DEFAULTS
-from daylio_to_md import utils, errors
-from daylio_to_md.entry.mood import Moodverse
+from obsidian_daylio_parser.config import DEFAULTS
+from obsidian_daylio_parser import utils, errors
+from obsidian_daylio_parser.entry.mood import Moodverse
 
 
 """---------------------------------------------------------------------------------------------------------------------
