+++ tests/test_librarian.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
 
-from daylio_to_md.entry.mood import Moodverse
-from daylio_to_md.journal_entry import EntryBuilder
-from daylio_to_md.librarian import Librarian, CannotAccessJournalError
-from daylio_to_md.group import EntriesFromBuilder
-from daylio_to_md.reader import CsvJournalReader
+from obsidian_daylio_parser.entry.mood import Moodverse
+from obsidian_daylio_parser.journal_entry import EntryBuilder
+from obsidian_daylio_parser.librarian import Librarian, CannotAccessJournalError
+from obsidian_daylio_parser.group import EntriesFromBuilder
+from obsidian_daylio_parser.reader import CsvJournalReader
 
 
 class TestLibrarian(TestCase):
